Consolidate interfaces for mockgen
